requirements.txt
langchain-core
langchain
langchain-openai
python-dotenv
pydantic